Develop

+++ src/database/queries/articles.query.js
@@ -4,9 +4,9 @@ SELECT
   a.title,
   a.description,
   a.slug,
-  a.price,
+  a.price::float,
   a.active,
-  a.create_at,
+  a.created_at,
   a.updated_at,
   json_build_object(
     'id', u.id::integer,
@@ -24,7 +24,7 @@ ON a.user_id = u.id
 LEFT JOIN images_article i
 ON a.id = i.article_id
 WHERE u.id = $1
-GROUP BY a.id, a.title, a.description, a.slug, a.price, a.active, a.create_at, a.updated_at, u.id, u.name, u.email, u.image
+GROUP BY a.id, a.title, a.description, a.slug, a.price, a.active, a.created_at, a.updated_at, u.id, u.name, u.email, u.image
 ORDER BY a.id
 OFFSET $2
 LIMIT $3
@@ -36,9 +36,9 @@ a.id::integer,
 a.title,
 a.description,
 a.slug,
-a.price,
+a.price::float,
 a.active,
-a.create_at,
+a.created_at,
 a.updated_at,
 json_build_object(
   'id', u.id::integer,
@@ -56,7 +56,7 @@ ON a.user_id = u.id
 LEFT JOIN images_article i
 ON a.id = i.article_id
 WHERE a.active = true
-GROUP BY a.id, a.title, a.description, a.slug, a.price, a.active, a.create_at, a.updated_at, u.id, u.name, u.email, u.image
+GROUP BY a.id, a.title, a.description, a.slug, a.price, a.active, a.created_at, a.updated_at, u.id, u.name, u.email, u.image
 ORDER BY a.id
 OFFSET $1
 LIMIT $2
@@ -68,9 +68,9 @@ a.id::integer,
 a.title,
 a.description,
 a.slug,
-a.price,
+a.price::float,
 a.active,
-a.create_at,
+a.created_at,
 a.updated_at,
 json_build_object(
   'id', u.id::integer,
@@ -87,7 +87,7 @@ INNER JOIN users u
 ON a.user_id = u.id
 LEFT JOIN images_article i
 ON a.id = i.article_id
-GROUP BY a.id, a.title, a.description, a.slug, a.price, a.active, a.create_at, a.updated_at, u.id, u.name, u.email, u.image
+GROUP BY a.id, a.title, a.description, a.slug, a.price, a.active, a.created_at, a.updated_at, u.id, u.name, u.email, u.image
 ORDER BY a.id
 OFFSET $1
 LIMIT $2
@@ -99,9 +99,9 @@ a.id::integer,
 a.title,
 a.description,
 a.slug,
-a.price,
+a.price::float,
 a.active,
-a.create_at,
+a.created_at,
 a.updated_at,
 json_build_object(
   'id', u.id::integer,
@@ -119,7 +119,7 @@ ON a.user_id = u.id
 LEFT JOIN images_article i
 ON a.id = i.article_id
 WHERE a.id = $1
-GROUP BY a.id, a.title, a.description, a.slug, a.price, a.active, a.create_at, a.updated_at, u.id, u.name, u.email, u.image
+GROUP BY a.id, a.title, a.description, a.slug, a.price, a.active, a.created_at, a.updated_at, u.id, u.name, u.email, u.image
 `;
 
 export const GET_ARTICLE_BY_SLUG = `
@@ -128,9 +128,9 @@ a.id::integer,
 a.title,
 a.description,
 a.slug,
-a.price,
+a.price::float,
 a.active,
-a.create_at,
+a.created_at,
 a.updated_at,
 json_build_object(
   'id', u.id::integer,
@@ -148,7 +148,7 @@ ON a.user_id = u.id
 LEFT JOIN images_article i
 ON a.id = i.article_id
 WHERE a.active = true AND a.slug = $1
-GROUP BY a.id, a.title, a.description, a.slug, a.price, a.active, a.create_at, a.updated_at, u.id, u.name, u.email, u.image
+GROUP BY a.id, a.title, a.description, a.slug, a.price, a.active, a.created_at, a.updated_at, u.id, u.name, u.email, u.image
 `;
 
 export const GET_TOTAL_ARTICLES = `
@@ -180,7 +180,7 @@ UPDATE articles SET active = $1 WHERE id = $2
 `;
 
 export const DELETE_ARTICLE_BY_ID = `
-DELETE FROM articles WHERE id = $1 RETURNING *, id::integer
+DELETE FROM articles WHERE id = $1 RETURNING *, id::integer, price::float, user_id::integer
 `;
 
 export const SET_IMAGES_TO_ARTICLE = `
@@ -189,5 +189,5 @@ VALUES ($1, $2) RETURNING id::integer, url_img
 `;
 
 export const DELETE_IMAGE_BY_ARTICLE_IMAGE_ID = `
-DELETE FROM images_article WHERE article_id = $1 AND id = $2 RETURNING *;
+DELETE FROM images_article WHERE article_id = $1 AND id = $2 RETURNING *, id::integer, article_id::integer
 `;
